+++ forward_checking.java
@@ -24,7 +24,7 @@ public class forward_checking {
     public static int colNum;
 
     //Node counter variable
-    public static int nodeCount;
+    public static long nodeCount;
 
     //Stores the current selected heuristic
     public static String heuristic;
